Assumer le score · un nombre entier de points, de la déclaration à l'écran
La valeur d'une conversion est un score, jamais un montant. Elle traversait
pourtant toute la chaîne en flottant · la colonne, les signatures, le calcul
des tunnels.

- la colonne `events.value` devient un entier signé, dans sa création, et le
  relevé du schéma suit ;
- `TrackedEvent::define()`, `Analytics::record()` et `Funnel::step()` prennent
  un entier, comme tout ce qui en découle jusqu'au total d'un tunnel · les
  taux de conversion restent des flottants ;
- l'entrée refuse un score décimal ou que la colonne ne peut pas tenir, et le
  collecteur l'écarte plutôt que de faire refuser tout le lot qui le porte ;
- les écrans disent « score » et « points », la notice aussi · l'exemple du
  README enseignait un montant, et le marketing promettait une valeur qu'il
  n'a jamais calculée ;
- la page de reprise dit aux projets existants de passer leurs scores en
  entiers, et ne prétend plus que le paquet exige InnoDB ;
- les commentaires d'essai ne citent plus de règle hors du dépôt.

+++ database/migrations/2026_06_30_000003_create_falcon_analytics_events_table.php
@@ -50,7 +50,10 @@ return new class extends Migration
             $table->string('target_selector', 255)->nullable();
             $table->string('target_text', 255)->nullable();
             $table->json('props')->nullable();
-            $table->decimal('value', 12, 2)->nullable();
+
+            // A score in points, never an amount: a whole number, signed so a
+            // penalty can be written down.
+            $table->integer('value')->nullable();
 
             $table->string('subject_type', 32)->nullable();
             $table->unsignedBigInteger('subject_id')->nullable();
